backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv==0.19.2
requests==2.26.0
python-jose==3.3.0
sqlalchemy==1.4.25